feat: add optional monitoring config (openobserve or sentry)
